Geofence support
Added support for tracking a user's enter/exit/dwell
events within a geofence centered around the AAP tower.
When a user enters the geofence, an event is written to
a firebase realtime database, as well as when they exit.
A notification is sent upon "dwelling" to remind user to
save their parking location.